requirements.txt
pandas>=2.3.0
tabulate>=0.9.0
pdfplumber>=0.11.0
pdfminer.six>=20231228
camelot-py>=1.0.9
opencv-python-headless>=4.8.0
pypdf>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.1
